Solved Day1 and probably Day2.
Tried a bunch of different solutions for getting part 2 to solve

BigIntegers => Well, it is really big numbers. CPU and Mem intensive
Tried finding square roots
JUmping into Modulus, bot not knowing what to Modulu on.

Read some clues later on the internet to figure out what is the interesting part.
I have for now not submitted the answer for part 2 since I "cheated" a bit.